+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/configuration/wrapper/ErrorResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import uk.gov.justice.laa.crime.dces.integration.model.exception.TraceData;
 
 @Data
 @Builder
