+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/model/drc/SendFdcFileDataToDrcRequest.java
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UpdateLogFdcRequest {
+public class SendFdcFileDataToDrcRequest {
     private Integer fdcId;
-    private String errorText;
 }
